Feat/information bug fix (#89)
* FavoriteController 미로그인 회원 좋아요 반영되지 않도록 수정

* Design:디자인변경

+++ src/main/java/com/ll/FlexGym/domain/Favorite/controller/FavoriteController.java
@@ -31,11 +31,15 @@ public class FavoriteController {
         // 예를 들어, MemberService, InformationService를 통해 가져올 수 있다고 가정한다.
         // 좋아요를 눌렀을 때 isFavorite을 통해 해당 유저의favorite이 이미 있는지
 
-        if(!favoriteService.isFavorite(member.getId(), infoId)){
-            log.info("isFavorite = {}", favoriteService.isFavorite(member.getId(), infoId));
-            favoriteService.addFavorite(member.getId(), infoId);
-        }else{
-            favoriteService.removeFavorite(member.getId(), infoId);
+        if(member==null){
+            return "redirect:/usr/information/info";
+        }else {
+            if (!favoriteService.isFavorite(member.getId(), infoId)) {
+                log.info("isFavorite = {}", favoriteService.isFavorite(member.getId(), infoId));
+                favoriteService.addFavorite(member.getId(), infoId);
+            } else {
+                favoriteService.removeFavorite(member.getId(), infoId);
+            }
         }
         return "redirect:/usr/information/info";
     }
